chore: shuffle some things around
